+++ dags/sql_scripts/src_loans.sql
@@ -1,10 +1,10 @@
 -- "loan_id" int;"account_id" string;"date" date yymmdd;"amount" float;"duration" int;"payments" float;"status" low cardinality string/single char
 CREATE TABLE IF NOT EXISTS src_loans (
-    loan_id Int32,
+    loan_id UInt64,
     account_id String,
     date Date,
     amount Float32,
-    duration Int32,
+    duration UInt64,
     payments Float32,
     status LowCardinality(String)
 ) ENGINE = MergeTree()
